# backend/src/swagger.yaml
openapi: 3.0.3
info:
  title: CapstoneHub API
  version: 1.0.0
paths:
...
  /api/projects:
    post:
      summary: Submit project proposal
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created project
...
  /api/admin/reports/projects.csv:
    get:
      summary: Export projects report as CSV
      security:
        - bearerAuth: []
      responses:
        "200":
          description: CSV report
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT